manifest: drop the --filesystem=xdg-config/autostart permission
Notes implements the autostart feature through a third-party library
called qautostart[1], which has been recently patched by me to also
work on Flatpak apps.

But for qautostart to work properly on Flatpak apps, the
'--filesystem=xdg-config/autostart permission' is mandatory, which in
return causes the recently-added Flathub linter[2] to complain, and
to not build the app.

Instead, the new recommended way to implement the autostart feature is
through a "Flatpak portal" (i.e., a DBus interface).[3]

So, since I have no plans to implement this myself and I also deem the
autostart feature non-critical, I'm removing it altogether.

Users are free to add this Flatpak permission themselves, using either
a graphical tool like Flatseal[4], or by running this command:

flatpak override --user --filesystem=xdg-config/autostart io.github.nuttyartist.notes

Just keep in mind that this will only work on Notes >2.0.0, which at
the time of writing, is still unreleased.

References:
[1] https://github.com/b00f/qautostart
[2] https://github.com/flathub/flatpak-builder-lint
[3] https://docs.flatpak.org/en/latest/portal-api-reference.html#gdbus-org.freedesktop.portal.Background
[4] https://flathub.org/apps/details/com.github.tchx84.Flatseal